"chore: repalce double click and middle click action, logBox helpBox update, remove issue fixed"
